requirements.txt
boto3==1.14.26
botocore==1.17.26
Django==3.0.8
django-debug-toolbar==2.2
django-markdown2==0.3.1
django-markdownx==3.0.1
django-multiselectfield==0.1.12
django-storages==1.9.1
django-widget-tweaks==1.4.1
docutils==0.16
jmespath==0.10.0
Markdown==3.2.2
markdown2==2.3.9
numpy==1.19.1
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.2.4
